Fix warnings
git-svn-id: http://jules.aerix.nl/svn/g-key/trunk@269 f5aa9b1c-f3d2-43f6-b11e-4419a1f85e2c

+++ ts3_settings.cpp
@@ -36,7 +36,7 @@ bool TS3Settings::GetValueForQuery(std::string query, std::string& result)
 {
 	// Prepare the statement
 	sqlite3_stmt* sql;
-	if(CheckAndHandle(sqlite3_prepare_v2(settings, query.c_str(), query.length(), &sql, NULL)))
+	if(CheckAndHandle(sqlite3_prepare_v2(settings, query.c_str(), (int)query.length(), &sql, NULL)))
 		return false;
 
 	// Get the value
